feat: parameterize system combinators with Into<EffectOut> piping
The system combinators only accepted pipework from systems that return
EffectOut. This makes it so they accept pipework from systems that
return types that can be converted into EffectOut, and perform that
conversion.

In the trivial case, this is a no-op.

The main type that implements conversion to EffectOut is all Effects.
So, now, users can go as far as to return -> impl Effect, and it will be
acceptable, as demonstrated in the rainbow-clear-color example.