feat: add `ResWith` effect for transforming a resource with a function (#2)
In addition to `-Put` effects, the mvp will include `-With` effects,
which use a "transformation" function to perform the update. This can be
useful whenever the desired update is dependent on the current value.
This will mostly see use on iterative system params like `Query`, but it
has its purposes in `Res` as well, in case the user is producing many
effects on the same resource in the same pipeline and wants those
effects to compound properly.

Unfortunately, it does require that `R` implements `Clone`, for the time
being. I looked for a way to update a `&mut T` with a `Fn(T) -> T`
safely, but didn't find anything.