test: create `OneWayFn` test utility for function generation with `proptest` (#6)
The resource effects tests developed a method for creating functions
with `proptest` to test with. Instead of creating completely arbitrary
functions, they generate a salt and use that in a hashing function to
create a somewhat-arbitrary one-way-function. This is easier than
actually generating an arbitrary function, and has a nice property of
guaranteeing that the correct function was run on the correct input when
testing code with higher-order functions.

However, there are a few issues with the current implementation. This
utility is very specific to the test resource `NumberResource`. More
effects will leverage higher-order functions, so it would be nice if it
was more generically useful. Furthermore, it doesn't "simplify" super
well with `proptest`'s shrinking strategies. Finally, it doesn't
actually provide a `proptest` strategy, i.e., there are extra steps to
generate the function using a salt as `proptest` input.

This change creates a `OneWayFn` test type that addresses all of these
shortcomings.