feat: add ComponentsPut and ComponentsWith effects (#7)
One of the most important side effects of bevy systems is updating the
state of components. This is mostly done using the `Query` system
parameter. This change introduces two effects for defining state
transitions of components in queries: `ComponentsPut` and
`ComponentsWith`:
- `ComponentsPut` sets a tuple of components to the provided values.
- `ComponentsWith` transforms a tuple of components with the provided
function. This function can also be parameterized by some
`ReadOnlyQueryData`.

These two apply their effects to every entity in a `Query`. They can be
defined with a `QueryFilter` to make the query more specific. Even more
specific entity-component effects will be provided in the future.