feat: add `EntityComponentsPut` and `EntityComponentsWith` effects (#11)
One of the most important side effects of bevy systems is updating the
state of components. Previously, `Components-` effects were added to
apply a component state transition to all entities in a query. This
change introduces two effects for definining state transitions of
components for a single entity:
- `EntityComponentsPut` sets a provided entity's tuple of components to
the provided values
- `EntityComponentsWith` transforms a provided entity's tuple of
components with the provided function. This function can also be
parameterized by some `ReadOnlyQueryData`.