docs: adjust mvp effects for bevy 0.16 design considerations (#15)
Some modern features of bevy need to be accounted for in the effects
mvp. Some of these may be older than bevy 0.16, but I was not familiar
with all of them as the last bevy version I had intense contact with was
0.12-ish.

In particular, running systems with commands, observers, relations, and
error handling are all pretty new. These are now accounted for in the
mvp checklist, with some things merely being "nice to have".

`Either` has been tossed in here as well, which should have probably
been included in the first place.