feat: add `CommandSpawnAnd` effect (#20)
One of the most basic ECS operations is the spawning of entities. This
is usually done with `Commands::spawn`, which queues up an entity to be
spawned with a provided set of components the next time the bevy
`apply_deferred` system is run.

However, it is common to use the resulting `Entity` id to cause more
effects, like creating relationships (especially hierarchies), storing
special entities in resources, or writing events regarding them. An
effect spawning an entity would not have access to its id on creation,
since it is spawned later during the `affect` system.

This change adds the `CommandSpawnAnd` resource. It can spawn entities
with bundles, but it can also cause more effects with the resulting
`Entity` by accepting closures like `|Entity| -> impl Effect`.