refactor: rename variadic effects to algebra effects (#22)
The planned `Either` effect is currently in the "iterator" category, but
it's really not an iterator. Really, it's a basic general-use sum type
for effects, whereas the existing tuple implementations are a more
complex general-use product type for effects. So, as far as
categorization goes, it can fit better in an "algebra" module than an
"iterator" module.