feat: add `IterEffect` effect and add `Vec` and `Option` implementations (#24)
There are times when you need to produce an effect a variable number of
times, or sometimes not at all. These situations are analogous to the
`Vec` and `Option` types, or more generally, iterators. This change adds
an `IterEffect` for generic iterators of effects, as well as
implementations for `Vec` and `Option`.