feat: add `AffectOrHandle` effect and `Result` implementation (#25)
The final important building block of effects is the `Result` effect.
Having this effect will be quite powerful, not only allowing users to
produce errors, but also partial failures if they desire (by using a
result in a nesting of effects). This change implements it with the help
of a new `AffectOrHandle` effect.