refactor: rename Put effects to Set, and With effects to SetWith (#27)
`Put` terminology is traditionally used for "update or insert" logic.
However, all of the existing `Put` effects are just "update", not
"insert". `Set` isn't quite as forgiving, and is more appropriate for
how these effects behave.

In a similar vein, `With` effects were just kind of confusing, `SetWith`
is a bit more clear.