refactor: standardize on structs instead of tuple-structs (#28)
There has been mixed usage of structs and tuple-structs among the
effects. Tuple structs tend to be less verbose to construct, but its
lack of field names can lead to unreadable code. Furthermore, the lack
of consistency throughout this library makes its interfaces
unpredictable, users will have to consult documentation or their LSP to
know what to type. This change makes it so everything public is a struct
with named fields for consistency and readability.