feat: add mapping methods for AffectOrHandle (#31)
The `Result` type has a lot of great declaritive/functional APIs. It
would be nice to be able to use these in a more declarative/functional
way when working with the `AffectOrHandle` type. This change provides
the `map_result` method, which will give you access to the entire
`Result` api. It also provides the `map` method, which is a more
convenient method for using the most common `Result` api of the same
name.