docs: make rainbow-clear-color example more basic (#35)
This repository needs a minimal example, and the goal of the
rainbow-clear-color example suits itself to this. Currently, it is more
complicated as it was used to try out some of the utility of the library
functions in-game. This change makes it a minimal example, and more
complex examples will be included later.