feat: upgrade to rust edition 2024 (#38)
Upgrading to the latest rust edition gets this crate in sync with the
rest of the rust community, especially the rest of the bevy ecosystem.
This change upgrades the project to rust edition 2024, updates
dependencies, and migrates the code (mostly adding `use<>` to systems
that return `-> impl Effect` and have some system parameters).