feat: add basic EffectOut composition methods (#43)
With the system combinators, `EffectOut` is something similar to a
monad, where the composition is applied between piped systems rather
than in simpler higher-order functions. This adds the simpler thing.
These will definitely be useful for composing `EffectOut`s within
systems if users decide to break down their logic into smaller,
effect-out-producing functions.

Namely, functor-ish and monad-ish methods that are available on `Result`
and `Option` are now also available for `EffectOut`:
- `flatten`
- `map`
- `and_then`

And then a couple others for more flexibility
- `flatten_compose`
- `and_then_compose`