feat: improve naming and documentation of system combinators (#44)
After writing the `EffectOut` composition methods, it became clear that
`and_compose` and `and_combine` were actually analogous to
`EffectOut::and_then_compose` and `EffectOut::and_then`. I think it
would be nice if their names were in homage as well. However, just
calling them module functions with those names and putting them in the
prelude would be presumptuous, especially a name such as common as
`and_then`. So, to individuate them and also express that they are
intended for pipe input, the `in_` prefix is used.

Documentation across the module has also been adjusted here to be more
similar to the standards set by the `EffectOut` changes.