docs: use Effect derive to simplify effect_composition doc tests (#45)
Before `Effect` was derivable, making a custom resource or message and
using `res_set` or `message_write` were the quickest ways to make dummy
effects for tests. The effect derive allows us to make `Effect` types
more directly, so now it can be used to simplify some of the code
examples. This changes the effect-composition module's doctests to use
simple `MyEffect` types.