feat: implement IntoIterator for AffectMany (#46)
`AffectMany` is just a wrapper around a generic `I: IntoIterator`. So,
it makes sense for it to also implement `IntoIterator`.

I considered also making it implement `FromIterator`, but this doesn't
seem as obviously good. In any situation where this would be ergonomic,
the target `I` parameter would still need to be supplied. So, I'm
leaving it be for now.