fix: move release-please config to config file (#52)
Initial runs of the release-please workflow result in some warnings
about deprecations and the use of certain config options in the job. It
turns out, release-please v4 requires many of the configuration options
to be specified in the config file itself. This change moves those
configurations to the config file, and also uses the newer location for
the release-please-action as the current one is deprecated.