feat: add book with outline (#53)
I would like to start on the right foot when it comes to documentation
of this library. This means having a place separate from the API
reference for writing tutorials, guides, and explanation docs. I've used
mdbook for this in the past to good effect, so I'm doing the same here.
This adds an mdbook to the repo with nothing but a simple outline.