feat: minimize bevy dependencies (#55)
This library currently just uses bevy's default dependencies, however we
do not need all of the defaults. Reducing our footprint on the bevy
feature flags will allow users that don't need all of the defaults to
compile quicker, and those who *can't* have all of the defaults are not
excluded. This change turns off default-features for bevy and minimizes
the set of feature flags used. In the future we may even want to reduce
our dependencies on these flags, or make them conditional as well.
`bevy_ui` and `bevy_camera` are only used in examples and doctests, so
these would be easy targets.