ci: add more changelog-sections for representing non-src changes in changelog (#59)
For the most part, we want the semver versions to respond to "fix" and
"feat" and "fix!" and "feat!" commit messages. However, there are many
changes that have nothing to do with the actual interface we're
versioning that I would still like to have represented in the changelog,
at least for recognition's sake when someone contributes. This adds some
changelog-sections configuration that should capture all the non-src,
non-crate changes in release-please.