book: write motivations chapter (#61)
I've been kind of wanting to write a bit about this library and my
thoughts behind it, and I think the book is an okay place to do it as it
may be interesting to users. It also kind of acts as a more long-winded,
stylistic explanation of the library.