example: add relationship example (#62)
One of the requirements when creating this library was being able to
spawn entity relationships in a single effect, where one entity's
component can store another entity. This might not be obvious at first
glance, so this provides an example demonstrating how to do it with a
simple transform hierarchy relationship. This code will likely be used
in a how-to-guide as well.