book: write 'Spawn a Relationship Synchronously' chapter (#64)
One of the most important requirements for this library was being able
to spawn relationships (especially hierarchys) synchronously, i.e., in
one system. We now have an example demonstrating how to do this, but I
wanted to tie this to some actual documentation as well. This creates a
how-to-guide for doing such a thing, using the example's code.