example: add sokoban example (#71)
It will be nice to have a slightly more complex example to demonstrate
the library being used for something closer to a real game, as well as
showing off more techniques. This adds a `sokoban` example, which takes
advantage of observer effects and the monadicness of `EffectOut`, among
other things.