book: share blurb between book and api ref (#76)
The README has a decent blurb briefly advertising the library. It would
be good to re-use this blurb in the book and api ref. In addition to
adjusting the blurb slightly, this copies it to a common file, and
includes it in both places. Note: it cannot be shared this way with the
README.md, we just need to make sure to update that manually.