book: describe documentation and add links in README.md (#80)
All documentation for this library should be discoverable. This adds
links to the book and api reference to the project README.md, which will
also be seen on crates.io. For now, the `main` branch of the book is
linked to, so that prospective users will actually be able to click this
link before release.