feat: reuse Result effect for EntityComponentsSet- error handling (#82)
The `EntityComponentsSet-` effects were written before the error effects
were. Their cop-out version of error handling was to just use a simple
`error!` log macro. This change makes it so they use a `Result<(),
QueryEntityError>` effect internally in the error case, which ultimately
makes them rely on the bevy `DefaultErrorHandler`.