feat: use no bevy features by default (#83)
As this is a library, we should do our best to minimize dependencies and
dependency features to avoid forcing them onto our users. The most
important of these for us is bevy. This change makes it so, by default,
we have no bevy features. The asset server effects need to be hid behind
a feature flag to make this possible. Various other changes were made to
simplify other targets dependencies (like examples, and the book).