fix: note that variadics_please and either are already included by bevy (#84)
I was originally planning to hide the Either effect and dependency
behind a feature flag. However, it turns out that, even with no default
features, bevy will bring in the either dependency. So, it's not really
worth it to make a feature flag out of it, which I've noted here.