fix: copy rainbow_clear_color example into README.md (#92)
The README is currently lacking in marketing material. Prospective users
should have a little bit of a better idea of what it's like to use the
library, or what it provides them, at a glance of the README. This
change just copies the rainbow_clear_color example into the README,
which is a minimal demonstration of the library at work.