example: add unit test to rainbow-clear-color example (#99)
I think one of the biggest values of having a system like this is the
ease of unit testing the pure system so a test case should be included.
I had to go through a bit of time and effort working out how to do it
without too much overhead. There may be easier ways to do this but this
is what I came up with.